Parse DefMethod
This just extracts the information needed to add it to the namespace and
progress with the parse. It does not actually parse the contents of the
method, as this only needs to be done if we want to execute it.

+++ aml_parser/src/pkg_length.rs
@@ -64,7 +64,7 @@ where
         }
 
         let (new_input, context, length): (&[u8], &mut AmlContext, u32) =
-            match take_n(byte_count as usize).parse(new_input, context) {
+            match take_n(byte_count as u32).parse(new_input, context) {
                 Ok((new_input, context, bytes)) => {
                     let initial_length = u32::from(lead_byte.get_bits(0..4));
                     (
